+++ src/parser/inline/emphasis.test.ts
@@ -22,6 +22,9 @@ describe('Unit: parser/inline/emphasis', () => {
       assert.deepStrictEqual(inspect(parser('*ab*')), [['<em>ab</em>'], '']);
       assert.deepStrictEqual(inspect(parser('*a\nb*')), [['<em>a b</em>'], '']);
       assert.deepStrictEqual(inspect(parser('*a\\\nb*')), [['<em>a<br>b</em>'], '']);
+      assert.deepStrictEqual(inspect(parser('*a**')), [['<em>a</em>'], '*']);
+      assert.deepStrictEqual(inspect(parser('*a**b*')), [['<em>a</em>', '<em>b</em>'], '']);
+      assert.deepStrictEqual(inspect(parser('*a**b**c*')), [['<em>a<strong>b</strong>c</em>'], '']);
     });
 
     it('nest', () => {