+++ src/parser/inline/emphasis.ts
@@ -1,18 +1,18 @@
 ﻿import { Result } from '../../combinator/parser';
 import { combine } from '../../combinator/combine';
 import { loop } from '../../combinator/loop';
-import { EmphasisParser, InlineParser, inline } from '../inline';
+import { EmphasisParser, StrongParser, InlineParser, inline } from '../inline';
 import { strong } from './strong';
 import { squash } from '../squash';
 
-type SubParsers = [InlineParser];
+type SubParsers = [InlineParser, StrongParser];
 
 const syntax = /^\*[\s\S]+?\*/;
 const closer = /^\*/;
 
 export const emphasis: EmphasisParser = function (source: string): Result<HTMLElement, SubParsers> {
   if (!source.startsWith('*') || source.search(syntax) !== 0) return;
-  const [cs, rest] = loop(combine<SubParsers, HTMLElement | Text>([inline]), rest => rest.slice(0, 99).search(closer) === 0 && !strong(rest))(source.slice(1)) || [[], ''];
+  const [cs, rest] = loop(combine<SubParsers, HTMLElement | Text>([loop(inline, closer), strong]))(source.slice(1)) || [[], ''];
   if (!rest.startsWith('*')) return;
   const el = document.createElement('em');
   void el.appendChild(squash(cs));